Move symbol hiding code to a new file
